Updating assessments/estimate-pi/EstimatePi.java

+++ assessments/estimate-pi/EstimatePi.java
@@ -68,7 +68,7 @@ public class EstimatePi implements ImageGenerator {
   // circle is πr^2 and the area of the square is (2r)^2 or 4r^2.
   // Therefore the proportion should be π/4 and thus π is 4 times the
   // proportion of darts that land in the circle.
-  public double estimate (double darts){
+  public double estimate (int darts){
     return darts*4;
   }
